Extract a slim viewport store from displayStore.ts, delete the rest
displayStore.ts (627 lines) mixed a live viewport-tracking concern
(viewportSize/isTouchDevice/isMobileViewport + resize-listener init,
used by kind-loader.vue and memoryStore.ts) with a dead
header/sidebar/footer display-mode state machine consumed only by five
components with no live render site outside WonderLab preview
fixtures (screen-debug, smart-icons, icon-display, smart-nav,
butterfly-toggle).

- New stores/viewportStore.ts carries just the viewport piece.
- kind-loader.vue and memoryStore.ts rewired to it.
- Deleted stores/displayStore.ts and the five dead components.
- Deleted stores/helpers/displayHelper.ts (displayStore's only
  consumer), stores/helpers/applyQuerySelections.ts and
  stores/seeds/modeLinks.ts (both zero-consumer dead code that only
  existed to feed displayStore's dead state machine), and the
  now-dangling loadDisplayStore export in stores/index.ts.

Conductor: interface-vision/t-045
