--- stores/helpers/displayHelper.ts
@@ -1,81 +1,0 @@
-// /stores/helpers/displayHelper.ts
-
-export function setCustomVh() {
-  if (typeof window !== 'undefined') {
-    const vh = window.innerHeight * 0.01
-    document.documentElement.style.setProperty('--vh', `${vh}px`)
-  }
-}
-
-export type NavDock = 'top' | 'bottom'
-export type SmartState = 'front' | 'back' | 'dash'
-
-export type DisplayState =
-  | 'hidden'
-  | 'compact'
-  | 'open'
-  | 'priority'
-  | 'disabled'
-
-export type FlipState = 'tutorial' | 'main' | 'toTutorial' | 'toMain'
-export type FullscreenState = 'nuxt' | 'fullscreen' | 'splash'
-
-export type displayModeState =
-  | 'scenario'
-  | 'character'
-  | 'reward'
-  | 'user'
-  | 'chat'
-  | 'bot'
-  | 'brainstorm'
-  | 'dream'
-  | 'art'
-  | 'collection'
-
-export type displayActionState =
-  | 'gallery'
-  | 'card'
-  | 'add'
-  | 'edit'
-  | 'generate'
-  | 'interact'
-  | 'dreammaker'
-  | 'brainstorm'
-
-export type EffectId =
-  | 'bubble-effect'
-  | 'fizzy-bubbles'
-  | 'rain-effect'
-  | 'butterfly-animation'
-
-export type ViewportSize = 'small' | 'medium' | 'large' | 'extraLarge'
-
-export interface DisplayStoreState {
-  headerState: DisplayState
-  sidebarLeftState: DisplayState
-  sidebarRightState: DisplayState
-  footerState: DisplayState
-  navDock: NavDock
-  isVertical: boolean
-  viewportSize: ViewportSize
-  isTouchDevice: boolean
-  showTutorial: boolean
-  isInitialized: boolean
-  flipState: FlipState
-  isFullScreen: boolean
-  isMobileViewport: boolean
-  isAnimating: boolean
-  currentAnimation: EffectId | ''
-  resizeTimeout: ReturnType<typeof setTimeout> | null
-  fullscreenState: FullscreenState
-  displayMode: displayModeState
-  displayAction: displayActionState
-  previousRoute: string
-  mainComponent: string
-  showLeft: boolean
-  showCenter: boolean
-  showRight: boolean
-  showExtended: boolean
-  showCorner: boolean
-  SmartState: SmartState
-}
